refactor(data_structures): decouple queue systems and demo coordinator to demos/data_structures
